Rename the slicer
