+++ src/components/complete/Complete.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useDispatch } from "react-redux";
 import { v4 as uuid } from 'uuid';
-import { addComplete } from "../../redux/features/counterSlice";
+import { addComplete } from "../../redux/features/todoSlice";
 import FormItem from "../todo/FormItem";
 
 const Complete = () => {
