+++ src/redux/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit'
-import todoReducer from './features/counterSlice'
+import todoReducer from './features/todoSlice'
 export const store = configureStore({
   reducer: {
     todo: todoReducer
